+++ pyls/config/source.py
@@ -29,3 +29,53 @@ class ConfigSource(object):
         for filename in files:
             found_files.extend(config.read(filename))
         return config
+
+    def parse_config(self, config, key, options):
+        """Parse the config with the given options."""
+        conf = {}
+        for source, destination, opt_type in options:
+            opt_value = _get_opt(config, key, source, opt_type)
+            if opt_value is not None:
+                _set_opt(conf, destination, opt_value)
+        return conf
+
+
+def _get_opt(config, key, option, opt_type):
+    """Get an option from a configparser with the given type."""
+    for opt_key in [option, option.replace('-', '_')]:
+        if not config.has_option(key, opt_key):
+            continue
+
+        if opt_type == bool:
+            return config.getbool(key, opt_key)
+
+        if opt_type == int:
+            return config.getint(key, opt_key)
+
+        if opt_type == str:
+            return config.get(key, opt_key)
+
+        if opt_type == list:
+            return _parse_list_opt(config.get(key, opt_key))
+
+        raise ValueError("Unknown option type: %s", opt_type)
+
+
+def _parse_list_opt(string):
+    return [s.strip() for s in string.split(",") if s.strip()]
+
+
+def _set_opt(config_dict, path, value):
+    """Set the value in the dictionary at the given path if the value is not None."""
+    if value is None:
+        return
+
+    if '.' not in path:
+        config_dict[path] = value
+        return
+
+    key, rest = path.split(".", 1)
+    if key not in config_dict:
+        config_dict[key] = {}
+
+    _set_opt(config_dict[key], rest, value)
